Emails: name the email plan in scheduled downgrade notices
The pending downgrade notice on the purchase page and the expiry status
both resolve the target product name with getPlanNames(), which only
covers site plans. An email tier fell back to a message with no product
name.

Adds getDowngradeTargetProductName(), which falls back to the email tier
name, and moves getTitanTierName() into the shared tier utils.
